requirements.txt
streamlit==1.22.0
openai==0.27.6
anthropic==0.5.0
python-dotenv==1.0.0
groq==0.9.0